CLI: fix Github reinstall

+++ out/repl.js
@@ -7580,35 +7580,28 @@ function printErrorAndExit(message) {
 async function dependencyResolve(dependency) {
   (0,external_fs_.mkdirSync)(".ix", { recursive: true });
   try {
-    switch (dependency.kind) {
-      case "http install": {
-        const response = await node_ponyfill_default()(dependency.url);
-        if (!response.ok) {
-          throw `${response.status}: ${response.statusText}: ${dependency.url}`;
-        }
-        const text = await response.text();
-        (0,external_fs_.writeFileSync)(`.ix/${dependency.alias}.ix`, text);
-        break;
+    if (dependency.kind === "http install") {
+      const response = await node_ponyfill_default()(dependency.url);
+      if (!response.ok) {
+        throw `${response.status}: ${response.statusText}: ${dependency.url}`;
       }
-      case "alias remove":
-        (0,external_fs_.unlinkSync)(`.ix/${dependency.alias}.ix`);
-        break;
-      case "Github install": {
-        const path = `.ix/${dependency.repo}`;
-        if ((0,external_fs_.existsSync)(path)) {
-          (0,external_fs_.unlinkSync)(`.ix/${dependency.repo}`);
-        }
-        await clone(`https://github.com/${dependency.repo}.git`, path, {
-          shallow: true
-        });
-        break;
-      }
-      case "Github remove":
+      const text = await response.text();
+      (0,external_fs_.writeFileSync)(`.ix/${dependency.alias}.ix`, text);
+    } else if (dependency.kind === "alias remove") {
+      (0,external_fs_.unlinkSync)(`.ix/${dependency.alias}.ix`);
+    } else if (dependency.kind === "Github install" || dependency.kind === "Github remove") {
+      const path = `.ix/${dependency.repo}`;
+      if ((0,external_fs_.existsSync)(path)) {
         (0,external_fs_.rmSync)(`.ix/${dependency.repo.split("/")[0]}`, {
           recursive: true,
           force: true
         });
-        break;
+      }
+      if (dependency.kind === "Github install") {
+        await clone(`https://github.com/${dependency.repo}.git`, path, {
+          shallow: true
+        });
+      }
     }
   } catch (err) {
     const message = typeof err === "object" && err && "message" in err ? err["message"] : `${err}`;
